refactor(core-api): restructure common and user domain directories
- Restructured shared components under the `common` package
- Moved user-specific exceptions into the `domain` layer
- Renamed `ui` to `presentation` to clearly represent external interface layer
