+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/common/support/error/CouponException.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.api.support.error;
+package dev.be.coupon.api.common.support.error;
 
 public class CouponException extends RuntimeException {
