+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/common/support/error/ErrorCode.java
@@ -1,0 +1,5 @@
+package dev.be.coupon.api.common.support.error;
+
+public enum ErrorCode {
+    E400, E500
+}
