+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/common/support/error/ErrorType.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.api.support.error;
+package dev.be.coupon.api.common.support.error;
 
 import org.springframework.boot.logging.LogLevel;
 import org.springframework.http.HttpStatus;
