+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/common/support/filter/v1/HttpLogMessage.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.api.support.filter.v1;
+package dev.be.coupon.api.common.support.filter.v1;
 
 import org.springframework.http.HttpStatus;
