--- coupon/coupon-api/src/main/java/dev/be/coupon/api/support/error/ErrorCode.java
@@ -1,5 +1,0 @@
-package dev.be.coupon.api.support.error;
-
-public enum ErrorCode {
-    E400, E500
-}
