--- coupon/coupon-api/src/main/java/dev/be/coupon/api/support/response/ResultType.java
@@ -1,5 +1,0 @@
-package dev.be.coupon.api.support.response;
-
-public enum ResultType {
-    SUCCESS, ERROR
-}
