+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/ui/HealthController.java
@@ -1,6 +1,6 @@
 package dev.be.coupon.api.ui;
 
-import dev.be.coupon.api.support.response.CommonResponse;
+import dev.be.coupon.api.common.support.response.CommonResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
