+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/user/presentation/UserControllerDocs.java
@@ -1,11 +1,9 @@
-package dev.be.coupon.api.user.ui;
+package dev.be.coupon.api.user.presentation;
 
-import dev.be.coupon.api.support.response.CommonResponse;
-import dev.be.coupon.api.user.ui.dto.UserSignUpRequest;
-import dev.be.coupon.api.user.ui.dto.UserSignUpResponse;
+import dev.be.coupon.api.common.support.response.CommonResponse;
+import dev.be.coupon.api.user.presentation.dto.UserSignUpRequest;
+import dev.be.coupon.api.user.presentation.dto.UserSignUpResponse;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
