+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/user/presentation/dto/UserSignUpRequest.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.api.user.ui.dto;
+package dev.be.coupon.api.user.presentation.dto;
 
 public record UserSignUpRequest(String username, String password) {
 }
